UI rework (#17)
* -Adjusted Actor and Component editing canvas size and position.

* Added BuilderStyles which allow to predetermine the size of buttons in UI element templates functions.

* GEE-8
-Added GEditorSettings and moved there booleans bDebugRenderComponents an bDebugRenderPhysicsMeshes.
-Editor settings are now serialized.
-Categorized settings in UI (there are three sections for now: editor, video, physics).

* Converting sln to CMake. I'm trying to make this multiplatform

* Working on the games pipeline so you can actually make them (#16)

* -Input box caret implemented. You can now paste text, enter it somewhere else than at the end, delete all text on the left or on the right of the caret.
-Slightly optimised basic shape rendering

* -Connecting to PVD during runtime
-Started adding physics properties to settings
-Editing collision shape material and collision object damping
-Button state now works using modifier bits rather than enums to avoid having to list out all the posibilities.
-Passing negative deltaTime to Interpolations does not allow the T value to go negative anymore, which lets you play animations in reverse

* -Adjusted the Event system so they are now grouped by Scene rather by window. You can now push an event to a single scene rather than to all scenes corresponding to a window. If you choose to push an event without specifying a scene, the active scene will be chosen.
-Fixed input box carets in canvases not resizing after zooming in/out of these canvases.

* -Small project restructuration
-More instantiation options will be added soon (instantiate to different actors)

* -Refactoring instantiation of hierarchy trees in progress

* -This version contains files and profiling features using nvapi from testing that I did of Deferred Shading against Forward Shading for my Extended Essay. The branch serves as an archive. Perhaps my future self will look at this code one day and laugh out loud. (#12)

-The features (mostly fixes) to be salvaged from this will hopefully be described in the master branch.


* Back after 6 months - finally done with exams! Development will continue and lots of things will be done this summer.
-PBR shader can now be forward rendered
-Added textbox caret
-Added the ability to customize max number of 2D and 3D shadow maps
-Simplified some naming

* In the process of fixing scrolling text components and merging the three text classes into one

* Fixed ScrollingTextComponent. It now renders correctly, but only if it is not in a UICanvas. This will be fixed in the next commit.
UIInputBoxActor Carets now work with scrolling text - the caret can move along with it, and it is positioned correctly when a scrolling text is clicked

* Scrolling text fixed for UICanvases. Carets are not hidden if they are outside input boxes, which must be fixed. Code will be cleaned up.

* Caret is not visible beyond input boxes anymore.
Cleaned up code.

* Moving the input box caret now scrolls text appropriately

* Caret position is normalised before using it in case it is set beyond the text string

* test commit

* Merged TextConstantSizeComponent with TextComponent. Tomorrow I will work on merging TextComponent with ScrollingTextComponent to make a versatile class with all 3 options combined

* Fixed a bug where scrolling text in UICanvases would not work properly when the canvas is moved from (0, 0). This was because in calculating scroll length we were converting from canvas space with FromCanvasSpace instead of ScaleFromCanvasSpace.
TextComponent and ScrollingTextComponent are still being merged.

* QoL: input boxes scroll back to 0 after pressing enter

* Fixed wrong window resolution read from tbcollection settings when rendering components in popup menus. This was caused by using the main tbcollection (with game settings, causing the bug) for popup rendering. Now each popup has its own tbcollection and settings can be read correctly.

* Compiled libraries for Win32 and x64 debug and release.
Updated assimp to 5.2.0
Updated glfw to 3.3.7 stable and OpenAL to kcat's 1.22.2
Updated freetype

* -Recast+Detour navmesh
-In-geditor grid and axis rendering
-Translate, rotate, resize actions in geditor


* -Converting to CMake + began using conan package manager instead of pushing external libs
-TextComponent render matrix transforms are now cached which significantly improves performance when multiple of them are visible. Recomputing them every frame used up 10-20% processing time.
-UI rework started which will allow to embed canvases into other canvases, procedurally generate their sizes, resize them etc. to produce a user-friendly layout.
-Fixed all warnings
-float time now converted to Time (which uses a double)

---------

+++ src/src/scene/hierarchy/HierarchyNodeInstantiation.cpp
@@ -52,7 +52,7 @@ namespace GEE
 					if (tree.GetAnimationCount() > 0)
 					{
 						AnimationManagerComponent& animManager = instantiationRoot.CreateComponent<AnimationManagerComponent>("An AnimationManagerComponent");
-						for (int i = 0; i < tree.GetAnimationCount(); i++)
+						for (int i = 0; i < static_cast<int>(tree.GetAnimationCount()); i++)
 							animManager.AddAnimationInstance(AnimationInstance(const_cast<Tree&>(tree).GetAnimation(i), instantiationRoot));
 					}
